Improve test coverage slightly
